// src/types/ProductTypes.ts
import { Seller } from "./SellerTypes";

export interface Product {
  id?: number;
  title: string;
  description: String;
  mrpPrice: number; // Manufacturer's Retail Price
  sellingPrice: number; // The price at which the product is sold
  discountPercentage: number; // Discount percentage applied to the product
  quantity: number; // Available quantity of the product
  color: string; // Color of the product
  images: string[]; // Array of image URLs for the product
  numRatings?: number; // Optional: Number of ratings received for the product
  category?: Category; // Optional: Category of the product
  seller?: Seller; // Optional: Seller of the product
  createdAt?: Date; // Optional: Date when the product was created
  sizes: string[]; // Sizes available for the product (e.g., "S,M,L" or "One Size")
}

// Interface for a Category object
export interface Category {
  id?: number; // Optional: Unique identifier for the category, since IDs are often auto-generated by the database
  name: string; // Name of the category (e.g., "Electronics", "Clothing")
  categoryId: string; // Unique string identifier for the category
  parentCategory?: Category; // Optional: Reference to the parent category for hierarchical categorization
  level: number; // The hierarchical level of the category (e.g., 0 for top-level, 1 for sub-category)
}

// Interface for a Seller object